Correct the header-includes processing
To preserve the type details of the 'header-includes' in the metadata,
we have to extract it from the metadata directly and via `get_metadata`.
If we don't, we only get strings which will not output correctly in the
LaTeX preamble.  This includes a test to show the correct behavior (and
it's how I found the bug).