Migrate to nox for testing
Instead of tox, we now use [nox][nox] to handle testing.  Nox provides a
bit better isolation of the environment than tox was doing (tox would
pass because it detected a personal installation of Pandoc).  We can use
the `conda` backend for the virtual environment so we can get a version
of Pandoc installed along with the panflute.  The other nicety of using
nox is the ability to add linting to the tests.

Note, the Python 3.6 test currently fails, but I'll be looking into that
soon.

[nox]: https://nox.thea.codes/en/stable/index.html