Add test demonstrating failed parsing around punctuation
Per #1, the parsing of the acronym keys fails when punctuation is next
to the key.  This appears to only happen when the acronym is not within
an explicit Span (i.e. inside '[+...]{}').  This test shows that the
issue is happening and gives us a place to test bed the solution.