Correct the parsing of acronyms with punctuation
An acronym can be at the end of a sentence or phrase.  To properly
handle this, we need to look for the longest possibly valid key and use
that.  We then take everything after that token as the punctuation or
potentially possessive marker and return that.  In the process, I
discovered that the quotes are slightly different.  Instead of being
strict punctuation, they are 'Quoted' blocks that need to be inspected
separately.