Check the punctuation returns the proper value
While we are checking that the key gets properly parsed, we also need to
check that the punctuation gets properly returned.  This is just one
more check in the test_get_key function.  But, we have to be careful
because we have to check if we are within a span and we should actually
have punctuation.  We also have to deal with the conversion of the
single quote to the fancy style.