Ensure the key is in the metadata before returning
Just because we found something that looks like a key does not mean it
is a key.  In particular, the [pandoc-xnos][xnos] family of filters also
use '+' to tag references.  We have to make sure the potential key is
actually in the acronyms metadata to say it is a key.

[xnos]: https://github.com/tomduck/pandoc-xnos