Strip the quotes from strings that are in Quoted blocks
It turns out that children of `panflute.Quoted` elements remember that
they are quoted.  That means they can remember that they need to be
quoted and subsequent calls to `panflute.stringify` include said quote.
However, it is not always the parent's quote type so we have to remove
all types of quotes.  For now, it appears like only straight quotes are
returned by `stringify`.  More tests to follow.