Fix the quotation tests
As stated in the documentation, the expected output when an acronym is
only used once is to _not_ include the acronym.  This could be added as
a feature, but I'll need to look into how `acro` handles it.  That means
we need to drop the short version in parenthesis for the tests.  This
also adds the tests that use fancy quotes.