Add mypy as a lint step
Flake8 does a good job checking the base syntax formatting, but mypy
adds type checking to the process.  This is a good thing.  It helps make
sure we get not only the annotations correct, but also have consistent
return values.  It also helps point out syntax that works, but may not
do what we expect in all cases (I'm looking at the string unpacking).