Move key parsing into the initialization
Instead of having the key only know about the base contents, it would be
better to have it fully parse the details.  This prevents us from having
to duplicate details in multiple locations.  For now, this boils down to
updating the `get` method to fully extract the key details without
changing the bare results.