Move plain text translation to the translate module
We are working on separating the core of the algorithm from the details
of the translation.  This means the plain text translation needs to move
to the translate module along with the LaTeX method.  So far, this is
just a cut and paste operation with a check that we still get the
correct result.