Correct the plain text starred output
Per the README, the correct behavior is expand every first usage as
`long (short)` until the first non-starred use and then as `short` after
that (unless the user said otherwise).  This does mean the original
tests become wrong so those had to be corrected.  While I was at it, the
proper plural full form doe not include the plural in the short version
of the acronym.