Add details on adding the list of acronyms
Now that we have a prototype, we can list out how to add the list of
acronyms to the document.  This also provides an explanation of what
options we will support and what the LaTeX output should be.  From
there, we can deduce what the plain text output should be.