Drop support for the heading option
It looks like dynamically adapting the heading option will be too
difficult at the outset.  Pandoc apparently does not like empty
attributes in divs.  That trips up the parsing leaving us with a
paragraph which is wrong.  That means the use _has_ to specify the
section or chapter which becomes pointless.  We may add the option back
later to pass to the acsetup macro.