Add the plain text acronym list generation
This adds the implementation of the acronym list generation for plain
text along with a test to make sure it works.  The basic logic is any
use of the short form of the acronym must trigger an entry in the
acronym list.  The initial determination of if an entry should be added
to the list is done by checking for the short form in the text; however,
this requires the acronym list to be placed _after_ the processing of
the rest of the document.