Establish the list when counting the usage
While we are counting the usage of the acronyms during initialization,
we can also determine if the acronym should be added to the list of
acronyms.  We know it will need to be added if the short or full version
was requested.  Additionally, the short version will be the default if
we have counted this key before and it has no explicit type.