Add a hook to build the docs when pushing to stable
A major point of working with this filter is to explore some new work
flows.  This script takes care of making sure the HTML documentation is
up to date by running Sphinx and adding the documents if any changes
occurred.  This only happens when the stable branch is being pushed and
it refuses to push if there are any lingering or staged changes.  This
is to prevent interfering with the user's state.  It will also checkout
the stable branch, do it's thing, and then change back to the original
branch.

To use this script, copy it to .git/hooks/pre-push and make it
executable.