Migrate the metadata to the setup.cfg
Instead of having a lot of static data in the setup.py, we can place it
in the setup.cfg as recommended by [PyPA] and [setuptools].  This file
is a bit easier to parse both visually and programmatically.  In fact,
we can now use the standard ConfigParser in the noxfile to extract the
dependencies.  This means we no longer have to update two lists if we
change the dependencies.

[PyPA]: https://packaging.python.org/tutorials/packaging-projects/#configuring-metadata
[setuptools]: https://setuptools.readthedocs.io/en/latest/userguide/quickstart.html