Fix the documentation build
Now that we have a separate function to setup the environment, we need
to use it.  However, I also discovered the paths were not quite right in
the sphinx command.  This is because I switched computers, but that
really means I need to update the command to get the paths correct.
While we're at it, we'll swap over to using pathlib instead of os for
path manipulations.