Move version to the setup.cfg
The better way to get the version for a module is through the
importlib.metadata.version function.  We don't really need an explicit
version in the source.  Instead, we place it with the rest of the
metadata and we'll be able to pull it in other contexts (like the sphinx
build).