Add a session to build the github docs
In the spirit of using a more portable tool, we can use nox directly to
build the html version of the documents for github.  We just run the
main docs session and then move the files into the correct location.
This is ultimately better than running make because we already know nox
is installed because we're testing.  It also prevents the headache of
not having the correct dependencies installed in a given environment
because it creates a fresh one for building.  The only additional change
is adding the githubpages extension to the sphinx build.