Update the MANIFEST.in
We don't want to include hidden files in the source distribution which
is what graft does.  Instead, we'll be explicit on what to include.
Fortunately, the [rules] for the MANIFEST.in state the tests will
already be included.  That means we only need to include the additional
data files.  Similarly for the documentation, we include all the
reStructuredText files.

[rules]: https://packaging.python.org/guides/using-manifest-in/