Add a post template with a by line
The rationale page needs an explicit by line to make it clear who the
author is.  To do this, we just need to adapt [this answer].  Now, a
page that sets the 'template' meta variable will redirect to the
appropriate template.  Thus, we just need to add the details for the by
line and set some minor customization.

[this answer]: https://stackoverflow.com/a/16368746/4249913