Pull supported Pythons for testing from setup.cfg
The setup.cfg lists the Pythons we claim to support.  Instead of hard
coding the versions in the noxfile.py, we can just pull them from
classifiers in the metadata.  This also helps with the fact that I
completely missed adding Python 3.9 to the supported list (in both
locations).  While we're at it, we can also pull the doc dependencies
from the pyproject.toml like we do for the install_requires.  It does
take some extra work to extract the dependencies because it's not a pure
INI file.