Add a running to do list
It's a good thing to keep a running to do list available.  This adds the
ability to list to dos anywhere in the project and it collects them into
a central list.  It also gives us a place to put some small
administrative to dos that don't really belong in the issue tracker.