Update the push hook to use nox
If we don't use the same version of Sphinx with which we built the
documents (as in, a lingering system version), we will still get a push
failure because Sphinx will generate a different set of documents.
Instead of calling `sphinx-build` separately, we can just rely on nox to
do the job.