Move the metadata state to an attribute of the document
Instead of munging with the metadata in the document, we can extract our
details and store them as an attribute on the document.  This make the
accessing syntax more Pythonic since we don't have to worry about
explicit type conversions (like integers being converted to strings).
This also gives us a single item to delete during the finalization step
so that we only change one thing.  This is mostly relevant if the filter
is used in tandem with others.  Previously, we were not cleaning up the
mess we were making.