Update the plain text replacement logic
In order to allow for different "full" usages, we have to do the full
usage determination a bit differently.  In particular, we have to query
the options for the appropriate form.  But, we also need to exercise
more caution when handling a single usage.  We add the logic to check
for the 'single' option and to select the appropriate form.  We also
drop the flag indicating usage and actually count the uses as we
traverse the document.  This does mean we need to rename the total usage
count in the acronym dictionary to reflect its real use.