Output the short option first in LaTeX headers
The documentation for acro states the short option _must be first_!.
It's an easy to overlook mistake that only requires us to explicitly
print short and then skip over it in the list comprehension.  However,
the test then becomes (actually always was) problematic.  We do not have
a guaranteed order we will get the keys.  Therefore, instead of doing
complex parsing, we'll just punt and force the keys to be in lexical
order (except for the necessary short).