Place the dependencies in the proper location
Per the [dependency management documentation][1], the proper place to
put additional dependencies is in the setup.cfg under the
'options.extras_require' field.  That lets us not need sphinx and the
styling at build time, but it also lets us fine tune the proper
installation details.  That does mean the noxfile needs to update, but
we can also use the [configuration parser that is in setuptools][2].
This prevents us from having to use a hand rolled version.  Note,
however, it does not parse the other sections so we still have to work
directly with the configparser when building the documents.

[1]: https://setuptools.pypa.io/en/latest/userguide/dependency_management.html
[2]: https://setuptools.pypa.io/en/latest/setuptools.html#configuration-api